backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "register-bot"
version = "0.1.0"
description = "Pipeline-based OpenAI registration assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.1",
  "pyyaml>=6.0.1",
  "httpx>=0.27.0",
  "cryptography>=42.0.0",
  "rich>=13.7.0",
  "click>=8.1.7",
  "structlog>=24.1.0",
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "PyJWT>=2.11.0",
  "patchright>=1.48.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
